Creato nuovo progetto GPUUpdaterPrj che è un nuovo exe che viene lanciato prima di GPU. Scopo di questo exe è quello di verificare la presenza di eventuali aggiornamenti di GPU nella cartella sul local FS autoUpdate; nel caso ci fossro aggiornamenti, li applica e poi lancia la GPU

+++ bin/aaa_MH6_CREATE_rheAPI.sh
@@ -18,6 +18,15 @@ if [ ! -f "$FILE_startRhea" ]; then
 	exit 1
 fi
 
+FILE_gpuUpdater="./EMBEDDED_RELEASE_GPUUpdater"
+if [ ! -f "$FILE_gpuUpdater" ]; then
+    echo "$FILE_gpuUpdater does not exist"
+	cp ./aaa_MH6_CREATE.sh ./error_EMBEDDED_RELEASE_GPUUpdate_does_not_exists
+	exit 1
+fi
+
+
+
 #la roba che mi interessa la metto tutta in GPUPackage2019
 rm -r ./GPUPackage2019
 mkdir ./GPUPackage2019
@@ -33,12 +42,13 @@ chmod 777 ./GPUPackage2019/makeRheaServicePack.sh
 chmod 777 ./GPUPackage2019/current/lang
 rm ./GPUPackage2019/varie/prog/js/dev/compiler.jar
 
-#fuori da package c'e' solo startRhea.sh e il miniboot
+#fuori da package c'e' solo startRhea.sh e gpuUpdater
 cp "$FILE_startRhea" ./
 chmod 777 ./startRhea.sh
 
+
 rm ./$filename
-tar -czvf $filename ./GPUPackage2019 ./startRhea.sh
+tar -czvf $filename ./GPUPackage2019 ./startRhea.sh ./"$FILE_gpuUpdater"
 
 rm ./startRhea.sh
 rm -r ./GPUPackage2019
